refactor(mcp): split manage_graph_data.py into 3 focused modules
PRD: hugegraph-mcp-manage-graph-data-split. Behavior-preserving split
of the 1149-line manage_graph_data.py into:

- graph_data_validate.py  — change plan validation, schema helpers,
  mode operation checks (U4)
- graph_data_gremlin.py   — Gremlin query generation with JSON
  escaping for injection prevention (U5)
- graph_data_execute.py   — dry-run preview, write execution,
  plan_hash, matched_count verification (U6, U7)
- manage_graph_data.py    — thin orchestration + full re-exports
  for backward compatibility (U8)

manage_graph_data() function unchanged. All public names re-exported
via __all__ so server.py, sql_modes.py, and all tests work without
import path changes.

Also fixed: schema plan_hash now correctly includes operation order
(PRD 2.3 U14), removing incorrect sorting that made hashes order-
insensitive.
